optimized_requirements.txt
# 优化后的依赖管理文件
# 核心依赖 - 精确版本控制
torch>=2.0.0,<2.2.0  # 限制版本范围，避免兼容性问题
transformers>=4.30.0,<4.40.0  # 限制版本范围
huggingface_hub>=0.16.0,<0.20.0
sentence-transformers>=2.2.0,<2.3.0
FlagEmbedding>=1.2.0,<1.3.0

# 数据处理
numpy>=1.21.0,<1.25.0
faiss-cpu>=1.7.0,<1.8.0  # 使用CPU版本，减少依赖
# faiss-gpu>=1.7.0,<1.8.0  # 如果需要GPU加速，取消注释

# 文本处理
PyPDF2>=3.0.0,<4.0.0
pdfplumber>=0.9.0,<1.0.0

# 系统监控
psutil>=5.9.0,<6.0.0

# 进度条和日志
tqdm>=4.64.0,<5.0.0

# 可选依赖（按需安装）
# accelerate>=0.20.0,<0.25.0  # 用于大模型加速
# safetensors>=0.3.0,<0.4.0  # 用于安全张量格式
# sentencepiece>=0.1.99,<0.2.0  # 用于某些分词器
# protobuf>=3.20.0,<4.0.0  # 用于某些模型配置

# 开发依赖（可选）
# pytest>=7.0.0,<8.0.0
# black>=22.0.0,<24.0.0
# flake8>=5.0.0,<7.0.0